Added a new "simulatePlatform" value which can be used to select which simulation mode the iTunes Emulator will run in. Setting it to "AppleTV" for example will expose the getSystemSounds object which can be used to generate navigation sounds.
Added support for "AppleTV" simulation. Some properties are NOT available on AppleTV and so are removed if this platform is selected; however there are also some features unique to the AppleTV. Mainly this is the SystemSounds object which allows you to generate navigation sounds.

Added navigations sounds for AppleTV simulation.